Ergonomics: Images group API is fully supported by builder pattern
