+++ async-openai/src/lib.rs
@@ -70,6 +70,6 @@ pub use edit::Edits;
 pub use embedding::Embeddings;
 pub use file::File;
 pub use fine_tune::FineTunes;
-pub use image::Image;
+pub use image::Images;
 pub use model::Models;
 pub use moderation::Moderation;
